Iterate on log level per computation unit
